pyproject.toml
[project]
name = "My AI Assistant"
version = "0.1.0"
description = "This project is mean to teach me more about how to use AI and how to build an AI Assistant. Specifically focussed on local agents for private use."
authors = [{name = "Faghmie Davids", email = "faghmie@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "dotenv >= 0.0.9",
    "python-dotenv >= 1.1.0",
    "ollama >= 0.4.8",
    "jira >= 3.8.0",
    "atlassian-python-api >= 4.1.3",
    "beautifulsoup4 >= 4.13.4",
    "html2text >= 2025.4.15",
    "slack_sdk >= 3.35.0",
]